pipelines/google_sheets/requirements.txt
python-dlt
google-api-python-client
google-auth
google_auth_oauthlib
